pyproject.toml
[project]
name = "knowledge-os"
version = "0.1.0"
description = "Spaced-repetition drilling + LLM-powered reading companion for self-study"
requires-python = ">=3.10"
dependencies = [
    "rich>=13.0.0",
    "readchar>=4.0.0",
    "textual>=0.47.0",
    # Reader module (Phase 1)
    "anthropic>=0.18.0",
    "google-genai>=1.0.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0",
    "jinja2>=3.0",
    "pymupdf>=1.23.0",
    "pymupdf4llm>=0.2.0",
    # Voice input (Phase 5)
    "sounddevice>=0.5.0",
    "faster-whisper>=1.0.0",
    "numpy>=1.24.0",
    "kokoro>=0.9.4",
    "soundfile>=0.13.1",
    "torch>=2.1",
    "spacy-lookups-data>=1.0.5",
    "nvidia-nccl-cu12>=2.27.5",
]

[tool.uv]
package = false
